third_party/WebKit/LayoutTests/animations/computed-style.html

<html>
<head>
<style type="text/css" media="screen">
    #test1 {
        width: 20px;
        height: 20px;
        background-color: blue;
        position: relative;
        animation-name: anim1;
        animation-duration: 10s;
        animation-fill-mode: backwards;
        animation-iteration-count: 10;
        animation-timing-function: linear;
        animation-direction: normal;
    }
    #test2 {
        width: 20px;
        height: 20px;
        background-color: blue;
        position: relative;
        animation-name: anim2, anim3;
        animation-duration: 5s, 2500ms;
        animation-fill-mode: forwards, both;
        animation-iteration-count: 10, infinite;
        animation-timing-function: linear, ease-in-out;
        animation-direction: normal, alternate;
    }
    #test3 {
        width: 20px;
        height: 20px;
        background-color: blue;
        position: relative;
        animation-name: anim1, anim2, anim3;
        animation-duration: 5s;
        animation-timing-function: step-start, step-end, steps(5, end);
    }
    @keyframes anim1 {
        from { left: 10px; }
        to { left: 20px; }
    }
    @keyframes anim2 {
        from { width: 20px; }
        to { width: 25px; }
    }
    @keyframes anim3 {
        from { left: 10px; }
        to { left: 20px; }
    }
</style>
<script src="../resources/js-test.js"></script>
</head>
<body>
<p id="test1"></p>
<p id="test2"></p>
<p id="test3"></p>
<p id="description"></p>
<div id="console"></div>
<script>

description("This tests computed style values from animation properties.");

var test1 = document.getElementById("test1");
var test2 = document.getElementById("test2");

var test1Style = window.getComputedStyle(test1);
var test2Style = window.getComputedStyle(test2);
var test3Style = window.getComputedStyle(test3);

shouldBe("test1Style.animationName", "'anim1'");
shouldBe("test2Style.animationName", "'anim2, anim3'");
shouldBe("test3Style.animationName", "'anim1, anim2, anim3'");

shouldBe("test1Style.animationDuration", "'10s'");
shouldBe("test2Style.animationDuration", "'5s, 2.5s'");

shouldBe("test1Style.animationFillMode", "'backwards'");
shouldBe("test2Style.animationFillMode", "'forwards, both'");

shouldBe("test1Style.animationIterationCount", "'10'");
shouldBe("test2Style.animationIterationCount", "'10, infinite'");

shouldBe("test1Style.animationTimingFunction", "'linear'");
shouldBe("test2Style.animationTimingFunction", "'linear, ease-in-out'");
shouldBe("test3Style.animationTimingFunction", "'step-start, step-end, steps(5, end)'");

shouldBe("test1Style.animationDirection", "'normal'");
shouldBe("test2Style.animationDirection", "'normal, alternate'");

debug("");

</script>
</body>
</html>
